feat(module): 完成二维地图功能的触发
1. 在原本的功能执行器Executor中分离出订阅器Subscriber，用于不同场景中监听功能模块列表的变化，激活对应功能
