+++ src/components/Map/Map2d/mixin/interaction/DrawingHelper/index.js
@@ -1,0 +1,14 @@
+/* 绘制帮助混入 */
+export default {
+  data() {
+    return {
+    };
+  },
+  methods: {
+  },
+  // 测距
+  measureDistance() {
+    // TODO: 完成测距逻辑
+    console.log(this.map2d);
+  },
+};
