+++ src/store/map2d/index.js
@@ -1,4 +1,4 @@
-import executor from './executor';
+import executor from '../executor';
 
 export default {
   namespaced: true,
